Added Component global
