+++ plugins/synth-react/src/components/SynthComponent/SynthComponent.tsx
@@ -23,30 +23,8 @@ import {
 } from '@backstage/core-components';
 import { Grid } from '@material-ui/core';
 import React, { useState } from 'react';
-import { usePlatformScript } from '../../hooks/usePlatformScript';
 import { YAMLEditor } from '../YAMLEditor/YAMLEditor';
-import { Inspector } from 'react-inspector';
-import { ps2js } from 'platformscript';
-
-const EvalResult = ({ yaml }: { yaml: string }) => {
-  const result = usePlatformScript(yaml || 'false');
-
-  if (result.loading) {
-    return <>Loading...</>;
-  } else if (result.error) {
-    return <>{result.error.message}</>;
-  } else if (result.value) {
-    if (result.value.type === 'external') {
-      if (React.isValidElement(result.value.value)) {
-        return result.value.value;
-      }
-      return <Inspector table={false} data={result.value.value} />;
-    }
-    return <Inspector table={false} data={ps2js(result.value)} />;
-  }
-
-  return <></>;
-};
+import { EvalResult } from './EvalResult';
 
 export const SynthComponent = ({ yaml }: { yaml: string }) => {
   const [_yaml, setYaml] = useState(yaml);
