--- plugins/synth-react/src/components/SynthComponent/myfile.yaml
@@ -1,1 +1,0 @@
-hello: world
